(BKR-1264) Ensure vm.hostname is legal
Some of the newer beaker-hosgenerator definitions contain underscores
which are illegal as hostnames, preventing vagrant configuring and
bringing up a machine (e.g. windows2012r2_ja-64-1). vmpooler for example
doesn't have a problem with this as the the actual hostname is obtained
from the vmpooler API, whereas beaker-vagrant re-uses the hostname
directly.

Translate any underscores in host.name to hyphens to ensure a legal
hostname.

rspec added to ensure translation happens.