BKR-1508 Add optional vagrant shell provision
Currently beaker does not support WinRM connections to windows hosts.
To use beaker to provision and run tests on a Windows 2016 server using vagrant,
beaker uses SSH connection to connect to the box. This means that there should be a SSH server
configured on the windows box. This needs to happen before beaker starting to connect.

The only possible option in the context of vagrant is to use vagrant provision.

So this add `shell_provisioner` option on the beaker host file to allow running
a shell script using vagrant. This takes advantage of Vagrant WinRM setup.