Work around common WinRM issue in Vagrant
Often, Windows systems will fail to shutdown or reboot cleanly when
using WinRM as a communications protocol.

After taking a look at the approach taken by Packer and Chef, it looks
like a simple retry system is the best mechanism for working around the
issue.

This patch retries up to 6 times with a 10 second delay between each in
the case of a WinRM error and just raises the traditional exception
otherwise.

A small amount of unnecessary code was removed.