Bugfix: Fix natdns logic
Before this patch, the (undocumented) Virtualbox host setting 'natdns'
would break Vagrant by adding the settings 'natdnshostresolver1' and
'natdnsproxy1' with the invalid valye "true", regardless of the text
'natdns' was set to for that host.  The correct value for these settings
is "on": https://www.virtualbox.org/manual/ch09.html#nat-adv-dns .

This patch fixes the settings by using the value "on" instead of "true".