Fix simultaneous runs with the libvirt hypervisor
Previously, if any two test runs occurred where the systems had identical
hostnames, then one of the runs would fail indicating that there was a
domain conflict.

This patch ensures that a unique name is created for each run when using
the `libvirt` provider.

Additionally, the string `beaker_` is prepended to the path name so that
it is easier to find your systems when large numbers of VMs are running
at the same time.

It also chooses a different base network when using `libvirt` vs
anything else so that users can use both VirtualBox and `libvirt`
simultaneously.