Correctly override the Vagrant path
In b162e77b2f8c8f465a6267385c5a7d3f1a2d2bda the path was updated to make
the path unique. However, it's done in the provisioning step. That's the
wrong time. If BEAKER_provision=no is passed, no provisioning happens
and the path is incorrect. By overriding the initializer, it always
happens.

Fixes: b162e77b2f8c8f465a6267385c5a7d3f1a2d2bda