Use Bundler to clean the environment
After updating to Fedora 38 (which includes Ruby 3.2 and Bundler 2.4) I
started to need this. Perhaps it's also a bug in the system wide vagrant
installation to not prune it, but Bundler provides a way to run with all
bundler-related variables removed. Since there was already some code to
try this, it now goes the whole way.