ssh.keep_alive: Add option to add the ssh keep alive option
We have an issue where sometimes the ssh connection is killed.  this
happens due to puppet changing the sshd keep alive settings while beaker
is still running.

This patch adds the ability for users to add ssh_keep_alive to the host
config to ensure we send keep alive packets every 5 seconds